Remove concatenated rickshaw JS with sources
* This would enable easier modification of functionality of individual
  modules. Having all of them in one single file makes it tough to track
  hierarchy of methods.

* As it is, there is no need to concatenate the files as not doing it
  won't affect the load times in any way. Another side effect of this
  would be that we can select only the modules we want to load.
