+++ web/index.html
@@ -7,7 +7,42 @@
     <script type="text/javascript" src="qrc:/web/src/objectcount.js"></script>
 
     <script type="text/javascript" src="qrc:/web/lib/d3.js"></script>
-    <script type="text/javascript" src="qrc:/web/lib/rickshaw.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Class.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Compat.ClassList.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Color.Palette.js"></script>
+
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Fixtures.Color.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Fixtures.Number.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Fixtures.RandomData.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Fixtures.Time.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Fixtures.Time.Local.js"></script>
+
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Ajax.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Annotate.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Axis.Time.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Axis.X.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Axis.Y.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Axis.Y.Scaled.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Behavior.Series.Highlight.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Behavior.Series.Order.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Behavior.Series.Toggle.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.HoverDetail.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.JSONP.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Legend.js"></script>
+
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Renderer.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Renderer.Area.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Renderer.Bar.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Renderer.LinePlot.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Renderer.Line.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Renderer.Multi.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Graph.Renderer.Stack.js"></script>
+
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Series.js"></script>
+    <script type="text/javascript" src="qrc:/web/lib/rickshaw/js/Rickshaw.Series.FixedDuration.js"></script>
+
     <link rel="stylesheet" href="qrc:/web/lib/rickshaw.css">
     <link rel="stylesheet" href="qrc:/web/src/rickchart.css">
   </head>
@@ -24,6 +59,7 @@
       </div>
     </div>
     <div id="legend"></div>
-    <script type="text/javascript" src="qrc:/web/src/rickchart.js"></script>
+
+    <script type="text/javascript" src="qrc:/web/src/rickchart.js"></script></script>
   </body>
 </html>
