+++ web/web.qrc
@@ -12,7 +12,6 @@
         <file>src/objectcount.coffee</file>
         <file>src/objectcount.js</file>
         <file>lib/d3.js</file>
-        <file>lib/rickshaw.js</file>
         <file>lib/rickshaw.css</file>
         <file>src/rickchart.coffee</file>
         <file>src/rickchart.css</file>
@@ -22,5 +21,42 @@
         <file>src/gcstat.js</file>
         <file>lib/exporting.js</file>
         <file>lib/highcharts-more.js</file>
+
+        <file>lib/rickshaw/js/Rickshaw.Class.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Color.Palette.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Compat.ClassList.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Fixtures.Color.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Fixtures.Number.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Fixtures.RandomData.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Fixtures.Time.Local.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Fixtures.Time.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Ajax.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Annotate.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Axis.Time.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Axis.X.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Axis.Y.Scaled.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Axis.Y.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Behavior.Series.Highlight.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Behavior.Series.Order.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Behavior.Series.Toggle.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.HoverDetail.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.JSONP.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Legend.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.RangeSlider.Preview.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.RangeSlider.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Renderer.Area.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Renderer.Bar.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Renderer.LinePlot.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Renderer.Line.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Renderer.Multi.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Renderer.ScatterPlot.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Renderer.Stack.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Renderer.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Smoother.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.Socketio.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Graph.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Series.FixedDuration.js</file>
+        <file>lib/rickshaw/js/Rickshaw.Series.js</file>
+        <file>lib/rickshaw/js/Rickshaw.js</file>
     </qresource>
 </RCC>
